Create Product of Array Except Self
Input: nums = [1, 2, 3, 4]

Output: [24, 12, 8, 6]

Explanation:

For output[0], the product of all elements except nums[0] is 2 * 3 * 4 = 24. For output[1], the product of all elements except nums[1] is 1 * 3 * 4 = 12. For output[2], the product of all elements except nums[2] is 1 * 2 * 4 = 8. For output[3], the product of all elements except nums[3] is 1 * 2 * 3 = 6.

Input Format

The input consists of a single line containing space-separated integers representing the elements of the array nums.

Constraints

2 <= nums.length <= 105
-30 <= nums[i] <= 30
Output Format

The output is an array of integers representing the output array.

Sample Input 0

1 2 3 4

Sample Output 0

24 12 8 6

Sample Input 1

-1 1 0 -3 3

Sample Output 1

0 0 9 0 0