Create Binary To Gray
The gray code is a binary numeral system where two successive values differ in only one bit. To convert a binary number to its corresponding gray code, follow these rules:

The Most Significant Bit (MSB) of the gray code is always equal to the MSB of the given binary code.
Other bits of the output gray code can be obtained by XORing the binary code bit at that index and the previous index.
Input Format

An integer 'n' representing the decimal value within the range of -10000 to 10000.

Constraints

-10000 to 10000

Output Format

An integer representing the equivalent gray code value of 'n'.

Sample Input 0

3

Sample Output 0

2

Explanation 0

In this case, the input value is 3. To convert it to gray code, we follow the rules mentioned in the problem description:

The Most Significant Bit (MSB) of the gray code is always equal to the MSB of the given binary code. Since the binary representation of 3 is 011, the MSB of the gray code will also be 0. For the other bits, we XOR the binary code bit at that index and the previous index. Starting from the rightmost bit, we have: The rightmost bit remains the same because there is no previous bit to XOR with. The middle bit is the result of XORing the current bit (1) with the previous bit (1), which gives us 0. The leftmost bit is the result of XORing the current bit (0) with the previous bit (1), which gives us 1. Thus, the gray code equivalent of the binary representation 011 (decimal 3) is 010 (decimal 2). Hence, the output is 2.

Sample Input 1

-11

Sample Output 1

15