Create Contiguous Sum Maximization
You are given an array of integers. Your task is to find the contiguous subarray with the maximum sum and return the sum.

Write a program that takes an array of integers as input and finds the maximum sum of a contiguous subarray.

Input Format

The first line of input contains an integer N representing the size of the array.
The second line of input contains N integers separated by a space representing the elements of the array.
Constraints

1 <= N <= 10^6
-10^9 <= Array[i] <= 10^9
Output Format

The program should output a single integer representing the maximum sum of a contiguous subarray.
Sample Input 0

9
-2 1 -3 4 -1 2 1 -5 4

Sample Output 0

6

Explanation 0

The subarray [4,-1,2,1] has the largest sum 6.

Sample Input 1

1
1

Sample Output 1

1
Explanation 1

The subarray [1] has the largest sum 1.